remove changes to other devices to minimise changes
